+++ company/Project/View/ProjectNoRoadCell.m
@@ -49,8 +49,8 @@
         UIButton *btn = (UIButton*)_btnArray[i];
         [btn setTitle:[NSString stringWithFormat:@"%@",model.areas[i]] forState:UIControlStateNormal];
     }
-    _personNumLabel.text = [NSString stringWithFormat:@"%ld",model.collectionCount];
-    _moneyLabel.text = [NSString stringWithFormat:@"%ld",model.financeTotal];
+    _personNumLabel.text = [NSString stringWithFormat:@"%ld",(long)model.collectionCount];
+    _moneyLabel.text = [NSString stringWithFormat:@"%ld",(long)model.financeTotal];
     
 }
 - (void)setSelected:(BOOL)selected animated:(BOOL)animated {
